fix issue with date immutability, break out mockito and H2 tests
